Optimize the return option of the iri:bao iri class
The return option of iri:bao now contains the explicit list of all
possible values. It allows express that a value of iri:bao is never
bao:BAO_0000015 that represents the bioassay type.

Because iri:bao can now represent only values valid for the endpoint
mapping, it has been moved to the appropriate endpoint files.